Update mail service
