+++ .env.example
@@ -1,0 +1,3 @@
+NEXT_PUBLIC_EMAILJS_SERVICE_ID = 
+NEXT_PUBLIC_EMAILJS_TEMPLATE_ID = 
+NEXT_PUBLIC_EMAILJS_PUBLIC_KEY =
