+++ app/components/homepage/contact/index.jsx
@@ -7,7 +7,7 @@ import { FaFacebook, FaStackOverflow } from 'react-icons/fa';
 import { FaXTwitter } from "react-icons/fa6";
 import { IoLogoGithub, IoMdCall } from "react-icons/io";
 import { MdAlternateEmail } from "react-icons/md";
-import { TbMailForward } from "react-icons/tb";
+import ContactForm from './contact-form';
 
 function ContactSection() {
   return (
@@ -19,47 +19,8 @@ function ContactSection() {
         <span className="h-36 w-[2px] bg-[#1a1443]"></span>
       </div>
       <div className="grid grid-cols-1 lg:grid-cols-2 gap-8 lg:gap-16 items-center">
-        <div className="">
-          <p className="font-medium mb-5 text-[#16f2b3] text-xl uppercase">
-            Contact with me
-          </p>
-          <div className="max-w-3xl text-[#d3d8e8] rounded-lg border border-[#464c6a] p-3 lg:p-5">
-            <p className="text-sm text-violet-200">
-              {"If you have any questions or concerns, please don't hesitate to contact me. I am open to any work opportunities that align with my skills and interests."}
-            </p>
-            <div className="mt-6 flex flex-col gap-4">
-              <div className="flex flex-col gap-2">
-                <label className="text-lg font-medium">Your email: </label>
-                <input
-                  className="bg-[#10172d] w-full border rounded-md border-[#353a52] focus:border-[#16f2b3] ring-0 outline-0 transition-all duration-300 px-3 py-2"
-                  type="email"
-                  maxLength="100"
-                  name="_replyto"
-                  required=""
-                />
-              </div>
-              <div className="flex flex-col gap-2">
-                <label className="text-lg font-medium">Your message: </label>
-                <textarea
-                  className="bg-[#10172d] w-full border rounded-md border-[#353a52] focus:border-[#16f2b3] ring-0 outline-0 transition-all duration-300 px-3 py-2"
-                  maxLength="500"
-                  name="message"
-                  required=""
-                  rows="5"
-                />
-              </div>
-              <div className="flex justify-center mt-3">
-                <button
-                  className="flex items-center gap-1 hover:gap-3 rounded-full bg-gradient-to-r from-pink-500 to-violet-600 px-5 md:px-12 py-2.5 md:py-3 text-center text-xs md:text-sm font-medium uppercase tracking-wider text-white no-underline transition-all duration-200 ease-out hover:text-white hover:no-underline md:font-semibold"
-                  role="button"
-                >
-                  <span>Send Message</span>
-                  <TbMailForward className="mt-1" size={18} />
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
+        <ContactForm />
+
         <div className="lg:w-3/4 ">
           <div className="flex flex-col gap-5 lg:gap-9">
             <p className="text-sm md:text-xl flex items-center gap-3">
