+++ app/components/homepage/experience/index.jsx
@@ -1,15 +1,15 @@
 // @flow strict
+
+import { experiences } from "@/utils/data/experience";
 import Image from "next/image";
-
-import { educations } from "@/utils/data/educations";
 import { BsPersonWorkspace } from "react-icons/bs";
-import AnimationLottie from "../helper/animation-lottie";
-import GlowCard from "../helper/glow-card";
-import lottieFile from '/public/lottie/study.json';
+import AnimationLottie from "../../helper/animation-lottie";
+import GlowCard from "../../helper/glow-card";
+import experience from '/public/lottie/code.json';
 
-function Education() {
+function Experience() {
   return (
-    <div id="education" className="relative z-50 border-t my-12 lg:my-24 border-[#25213b]">
+    <div id="experience" className="relative z-50 border-t my-12 lg:my-24 border-[#25213b]">
       <Image
         src="/section.svg"
         alt="Hero"
@@ -17,17 +17,12 @@ function Education() {
         height={795}
         className="absolute top-0 -z-10"
       />
-      <div className="flex justify-center -translate-y-[1px]">
-        <div className="w-3/4">
-          <div className="h-[1px] bg-gradient-to-r from-transparent via-violet-500 to-transparent  w-full" />
-        </div>
-      </div>
 
       <div className="flex justify-center my-5 lg:py-8">
         <div className="flex  items-center">
           <span className="w-24 h-[2px] bg-[#1a1443]"></span>
           <span className="bg-[#1a1443] w-fit text-white p-2 px-5 text-xl rounded-md">
-            Educations
+            Experiences
           </span>
           <span className="w-24 h-[2px] bg-[#1a1443]"></span>
         </div>
@@ -36,16 +31,16 @@ function Education() {
       <div className="py-8">
         <div className="grid grid-cols-1 lg:grid-cols-2 gap-8 lg:gap-16">
           <div className="flex justify-center items-start">
-            <div className="w-3/4 h-3/4">
-              <AnimationLottie animationPath={lottieFile} />
+            <div className="w-full h-full">
+              <AnimationLottie animationPath={experience} />
             </div>
           </div>
 
           <div>
             <div className="flex flex-col gap-6">
               {
-                educations.map(education => (
-                  <GlowCard key={education.id} identifier={`education-${education.id}`}>
+                experiences.map(experience => (
+                  <GlowCard key={experience.id} identifier={`experience-${experience.id}`}>
                     <div className="p-3 relative">
                       <Image
                         src="/blur-23.svg"
@@ -56,7 +51,7 @@ function Education() {
                       />
                       <div className="flex justify-center">
                         <p className="text-xs sm:text-sm text-[#16f2b3]">
-                          {education.duration}
+                          {experience.duration}
                         </p>
                       </div>
                       <div className="flex items-center gap-x-8 px-3 py-5">
@@ -64,10 +59,12 @@ function Education() {
                           <BsPersonWorkspace size={36} />
                         </div>
                         <div>
-                          <p className="text-base sm:text-xl mb-2 font-medium  uppercase">
-                            {education.title}
+                          <p className="text-base sm:text-xl mb-2 font-medium uppercase">
+                            {experience.title}
                           </p>
-                          <p className="text-sm sm:text-base">{education.institution}</p>
+                          <p className="text-sm sm:text-base">
+                            {experience.company}
+                          </p>
                         </div>
                       </div>
                     </div>
@@ -82,4 +79,4 @@ function Education() {
   );
 };
 
-export default Education;
+export default Experience;
