+++ app/page.js
@@ -1,7 +1,7 @@
 import { personalData } from "@/utils/data/personal-data";
-import AboutSection from "./components/homepage/about-section";
+import AboutSection from "./components/homepage/about";
 import Blog from "./components/homepage/blog";
-import ContactSection from "./components/homepage/contact-section";
+import ContactSection from "./components/homepage/contact";
 import Education from "./components/homepage/education";
 import Experience from "./components/homepage/experience";
 import HeroSection from "./components/homepage/hero-section";
