+++ package.json
@@ -9,12 +9,14 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@emailjs/browser": "^4.3.1",
     "lottie-react": "^2.4.0",
     "next": "latest",
     "react": "latest",
     "react-dom": "latest",
     "react-fast-marquee": "^1.6.2",
-    "react-icons": "^4.11.0"
+    "react-icons": "^4.11.0",
+    "react-toastify": "^10.0.4"
   },
   "devDependencies": {
     "autoprefixer": "latest",
