+++ utils/check-email.js
@@ -1,0 +1,5 @@
+export function isValidEmail(email) {
+  // Regular expression for validating email addresses
+  const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+  return emailRegex.test(email);
+};
